backend/requirements.txt
shioaji
fastapi
uvicorn[standard]
pandas
python-dotenv
python-dateutil
